upick/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
crispy-bootstrap5==2024.10
cssselect2==0.7.0
Django==5.1.6
django-crispy-forms==2.3
flup6==1.1.1
fonttools==4.56.0
Markdown==3.7
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
reportlab==4.3.1
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
weasyprint==64.1
webencodings==0.5.1
zopfli==0.2.3.post1
